refactor(error-logs): _utils 배럴 구조 정리 및 TSDoc 작성 규칙 문서화
getLogSummaryData 구현을 index.ts(배럴)에서 ErrorLogUtils.ts로 분리하고
TSDoc을 작성. 페이지별 _utils에 참고할 분리 사례가 없어 전역 utils의
PascalCase+Utils suffix 네이밍을 따름.

monitor-web CLAUDE.md에 _types/_utils의 index.ts는 re-export 전용이라는
규칙과, TSDoc은 사용자가 지목한 파일에만 요청 시 작성한다는 규칙을 추가.
